refactor(configmgr): move ADO system-git fallback code to git_ado.go
Isolates all Azure DevOps-specific code (IsAzureDevOpsURL,
cloneWithSystemGit, fetchWithSystemGit, writeAskpassScript, shellescape)
into a single file so it can be deleted in one shot when go-git v6
lands. The three inline branches in Start/schedule/Stop remain in
git.go as they are minimal call sites.
